CLOUDP-278034 - Missing replication test for secrets and certs  (#3838)
# Summary

PR adds missing `TestReconcileMultiClusterShardedClusterWithCustomTLS`
that verifies all needed certs and secrets are replicated in all
clusters

- simplified `ensureX509SecretAndCheckTLSType` function
- simplified `EnsureAgentKeySecretExists` function
- bumped `golangci-lint` to 1.61.0
- due to govet being more strict for non-constant format string there
are lots of changes like `workflow.Pending("%s", msg)`
- added
[comment](https://github.com/10gen/ops-manager-kubernetes/pull/3838/files#r1792125473)
for fishy function, not sure if it is correct
- some other small improvements

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.

+++ api/v1/om/opsmanager_types.go
@@ -869,7 +869,7 @@ func (om *MongoDBOpsManager) BackupDaemonStatefulSetNameForClusterIndex(clusterI
 	return fmt.Sprintf("%s-%d-backup-daemon", om.GetName(), clusterIndex)
 }
 
-func (om *MongoDBOpsManager) GetSchemePort() (corev1.URIScheme, int) {
+func (om *MongoDBOpsManager) GetSchemePort() (corev1.URIScheme, int32) {
 	if om.IsTLSEnabled() {
 		return SchemePortFromAnnotation("https")
 	}
@@ -1041,12 +1041,12 @@ func ConvertNameToEnvVarFormat(propertyFormat string) string {
 	return strings.Replace(withPrefix, ".", "_", -1)
 }
 
-func SchemePortFromAnnotation(annotation string) (corev1.URIScheme, int) {
+func SchemePortFromAnnotation(annotation string) (corev1.URIScheme, int32) {
 	scheme := corev1.URISchemeHTTP
-	port := util.OpsManagerDefaultPortHTTP
+	port := int32(util.OpsManagerDefaultPortHTTP)
 	if strings.ToUpper(annotation) == "HTTPS" {
 		scheme = corev1.URISchemeHTTPS
-		port = util.OpsManagerDefaultPortHTTPS
+		port = int32(util.OpsManagerDefaultPortHTTPS)
 	}
 
 	return scheme, port
